// src/app/users/enums/permissioncategories.enum.ts
export enum PERMISSION_CATEGORIES {
    QUOTATIONS = "quotations",            // Devis
    PROJECTS = "projects",                // Projets
    PRICE_REQUESTS = "price-requests",    // Demandes de prix
    PURCHASE_ORDERS = "purchase-orders",  // Bons de commande
    SUPPLIERS = "suppliers",              // Fournisseurs
    CATALOG = "catalog",                  // Catalogue
    USERS = "users",                      // Utilisateurs
    SEE_PRICES = "see-prices",            // Voir les prix
}